Cloud Removal with the Sentinel-2 L1C cloud mask

This function uses the fairly simple cloud mask which is shipped with the L1C data sets to remove cloudy pixels. A buffer can be applied to account for mixed pixels at the cloud edges as well as cloud shadow to some extent. The buffer assumes the projection of the data to be in meters, which is the case when the workflow of this package is used since the data is projected to the associated UTM zone.

cloudRemovalS2(
  datadir = "/home/rstudio/sen2tool/data/testregion/L1C",
  tifdir = "/home/rstudio/sen2tool/data/testregion/tifs",
  outdir = "/home/rstudio/sen2tool/data/testregion/clouds",
  tiffiles,
  safedirs,
  aoi,
  buffer = 100
)

Arguments

datadir
A charachter vector representing an absolute path to the original SAFE directories. From these directories the cloud mask is extracted. Note, that the naming conventions between the original SAFE products and the translated tif files differ slightly. Therefore, users are advised

tifdir
A charachter vector representing an absolute path to a directory where the translated tif files are found.

outdir
A charachter vector representing an absolute path to a directory where the cloud removed tif files are written to.

tiffiles
A charachter vector of variable length, usually representing the output of transTifS2. It indicates the absolute path of all files subject to the cloud removal process. safedirs for more very important information.

safedirs
A charachter vector of variable length, usually representing the output of downloadS2. Its elements have to correspond to the files specified in tiffiles. Note that the naming convention between the SAFE directories and the processed tiffiles differ slightly. Users are advised to only use the outputs of the other functions as inputs to cloudRemovalS2, because this way it is ensured that the vectors correspond to each other. This function does not include a check whether tiffiles and safedirs correspond. Thus the use of list.files() at the associated directories is strongly discouraged.

aoi
An object to which's extent the files are cropped.

buffer
A numeric vector of length 1 indicating the size of the buffer to draw around clouds specified in map units. When using the workflow of this package the unit is usually in meters since the data is projected to the corresponing UTM zone when transTifS2 is called.

Value

A charachter vector of variable length with the absolute path of the cloud removed tif files in outdir corresponding to the files specified in tiffiles.

Note

This function uses a system call to 'gdal_calc.py' to speed up the processs of cloud removal.