Evaluate deveolpment of a VI before and after an intervention year

This function takes calculated files of a given vegeatation index and evaluates its development before and after an intervention year summarised per polygon specifed in the aoi object. The metrics calculated are the mean, median, maximum, standard deviation and the number of observations. Additionally, the difference between the two epochs are calculated. When desired, the resulting rasters can be written to a directory.

evalVI(
  vifiles,
  pre_years,
  post_years,
  ncores = 1,
  aoi,
  name = "",
  returnRaster = F,
  rasterdir = NULL,
  extent_name = ""
)

Arguments

vifiles
A charachter vector indicating the file paths to the the vegetation index files.

pre_years
A charachter vector indicating the years of the pre-intervention epoch in the form "YYYY".

post_years
A charachter vector indicating the years of the post-intervention epoch in the form "YYYY".

ncores
The number of course to use for parallel processing. Defaults to 1, meaning the calculations are done in sequential mode.

aoi
An sf object with the polygons for which the metrics are averaged.

name
A charachter vector indicating the index represented by the raster layers. It will be appended to the resulting column names.

returnRaster
A logical indicating if the rasters should be written to disk. Defaults to FALSE.

rasterdir
Only relevant when returnRaster = TRUE. Otherwise a charchter vector pointing to an existing directory where the raster files are written to.

extent_name
A charachter vector indicating the extent name. It is going to be appended to the raster filenames when returnRaster = TRUE.

Value

The aoi sf object ammended by the statistcs of the pre and post epoch as well as calculated differences in the form of post - pre averaged per polygon. When returnRaster = TRUE the rasters are written to disk.